refactor: decouple td_cli.py, unify dry-run guards, and fix CI failures
- Extracted PR review submission logic into `dev-tools/submit_review.py`.
- Consolidated shared CLI logic into `dev-tools/utils.py`.
- Removed deprecated `fetch-review` alias.
- Implemented consistent `--dry-run` (default True) and `--execute` flags for all mutating commands.
- Updated documentation in `AGENTS.md` and agent workflows.
- Relocated unit tests to `tests/dev-tools/` and integrated them into `ci.yml`.
- Fixed CI Knip failures by removing unused dependencies `vite-plugin-pwa` and `workbox-window`.
